pyproject.toml
[build-system]
requires = ["setuptools>=77"]
build-backend = "setuptools.build_meta"

[project]
name = "endpoint-audit-demo"
version = "0.1.0"
description = "Consent-based endpoint auditing agent + ingestion server (no keystroke/clipboard capture)."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "ducktv1203" }]

dependencies = []

[project.optional-dependencies]
agent = [
  "httpx>=0.27",
  "psutil>=6.0",
  "pydantic>=2.7",
  "python-dotenv>=1.0",
]
server = [
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "pydantic>=2.7",
  "python-dotenv>=1.0",
]
dev = [
  "ruff>=0.6",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.setuptools.packages.find]
where = ["."]
include = ["agent*", "server*"]
exclude = ["storage*", "server_storage*", ".spool*"]

